tslint refactor
