+++ aop-navigation/.prettierrc
@@ -1,0 +1,5 @@
+{
+  "printWidth": 120,
+  "trailingComma": "all",
+  "singleQuote": true
+}
