+++ aop-navigation/projects/aop-routing/src/lib/model/models.ts
@@ -14,6 +14,7 @@ export interface AopNav {
 export interface AopNavigator {
     destinationPage?: string | number;
     navigationExtra?: NavigationExtras | undefined;
+    // tslint:disable-next-line:ban-types
     preprocess?: Function;
     param?: any;
 }
@@ -22,27 +23,8 @@ export class AopConfig {
     expirementNav: boolean;
 }
 
-/**
- * Class to encapsulate extra navigation properties or methods that's required to execute the navigation process.
- */
-export class NavAux {
-    constructor(public destinationPage?: string | number,
-         public navigationExtra?: NavigationExtras | undefined,
-         public preprocess?: Function, public param?: any) {
-    }
-}
 
-export abstract class AopBaseNavigation {
 
-    /**
-     * Skeleton methods to be overrided and provide custom navigation logic.
-     * Recommended to utilize NavAux class, as this class encapsulates navigation properties and will be passed
-     * by the decorator functions.
-     * @param navObj - Encapsulates navigation properties
-     */
-    abstract goToNextPage(navObj: NavAux, ...args): void;
-    abstract goToPreviousPage(navObj: NavAux, ...args): void;
-    abstract goToState(navObj: NavAux, ...args): void;
-}
+
